add menu to mobile popup
