+++ src/components/Pages/Guide/GuideMenu.js
@@ -1,0 +1,86 @@
+import React, { Component } from 'react';
+
+import classNames from 'classnames';
+import {hyphenate} from "./helpers";
+import { isMobileOrTablet } from 'js/helpers/reactMediaQueries';
+
+function GuideMenuLink({ title, anchorTag, isHeading, isCurrentSection }) {
+  // Each GuideSectionWrapper has an id={this.props.anchorTag}
+  function goToSection(e) {
+    e.preventDefault();
+    window.location.href = `#${anchorTag}`;
+  }
+
+  return (
+    <div
+      className={classNames("coa-GuideMenu__link", {
+        'coa-GuideMenu__heading': isHeading,
+        'coa-GuideMenu__subheading': !isHeading,
+        'coa-GuideMenu__current-section': isCurrentSection,
+      })}
+      onClick={goToSection}
+    >
+      {title}
+    </div>
+  )
+}
+
+function GuideMenuSection({ section, currentSection }) {
+  const headingAnchorTag = hyphenate(section.heading);
+  const subHeadings = section.pages.map((page, index) => {
+    const title = (
+      (page.servicePage && page.servicePage.title) ||
+      (page.informationPage && page.informationPage.title)
+    );
+    const anchorTag = `${headingAnchorTag}-${index+1}`;
+    return {
+      title,
+      anchorTag
+    }
+  });
+
+  return (
+    <div className="coa-GuideMenu__section">
+      <GuideMenuLink
+        title={section.heading}
+        anchorTag={headingAnchorTag}
+        isHeading={true}
+        isCurrentSection={currentSection === headingAnchorTag}
+      />
+      {subHeadings.map((subHeading, index) => (
+        <GuideMenuLink
+          title={subHeading.title}
+          anchorTag={subHeading.anchorTag}
+          isHeading={false}
+          isCurrentSection={currentSection === subHeading.anchorTag}
+        />
+      ))}
+    </div>
+  )
+}
+
+function GuideMenu({ contact, sections, currentSection }) {
+  return (
+    <div>
+      <div className="coa-GuideMenu__section">
+        {contact && (
+          <GuideMenuLink
+            title="Contact information"
+            anchorTag="Contact-information"
+            isHeading={true}
+            isCurrentSection={currentSection === "Contact-information"}
+          />
+        )}
+      </div>
+      {sections.map((section, index) => (
+        <GuideMenuSection
+          section={section}
+          currentSection={currentSection}
+          key={index}
+        />
+      ))}
+    </div>
+  )
+}
+
+export default GuideMenu;
