--- src/components/Pages/Guide/GuideSidebarMobile.js
@@ -1,52 +1,0 @@
-import React, { useState, useEffect, useMemo } from 'react';
-import classNames from 'classnames';
-
-import {hyphenate} from "./helpers";
-import dickens from "./dickens";
-
-function GuideSidebarMobile({ title, contact, sections, currentSection }) {
-  const [menuOpened, setMenuOpened] = useState(false);
-
-  // Freeze the body of the page when mobile menu is opened.
-  // Unfreeze the body of the page when mobile menu is closed.
-  const updateBodyStyling = () => {
-    if (menuOpened) {
-      window.document.body.classList.add("frozen-body");
-    } else {
-      window.document.body.classList.remove("frozen-body");
-    }
-  }
-  useMemo(updateBodyStyling, [menuOpened])
-
-  function toggleMenuOpened() {
-    setMenuOpened(!menuOpened)
-  }
-
-  const arrowIcon = (menuOpened) ? "expand_more" : "expand_less";
-
-  return (
-    <div className="coa-GuideSidebar__mobile-menu-container" onClick={toggleMenuOpened}>
-      <div className={classNames("coa-GuideSidebar__mobile-menu-popup", {
-        "coa-GuideSidebar__mobile-menu-popup-open": menuOpened
-      })}>
-        <div className="coa-GuideSidebar__mobile-menu-background"/>
-        <div className="coa-GuideSidebar__mobile-menu-content-container">
-          <div className="coa-GuideSidebar__mobile-menu-content-title">
-            <h2>Table of contents</h2>
-            <h3>{title}</h3>
-          </div>
-          <div>
-            {dickens}
-          </div>
-        </div>
-      </div>
-      <div className="coa-GuideSidebar__mobile-menu-button">
-        <i className="material-icons">list</i>
-        <span>Table of contents</span>
-        <i className="material-icons">{arrowIcon}</i>
-      </div>
-    </div>
-  )
-}
-
-export default GuideSidebarMobile;
