--- src/components/Pages/Guide/dickens.js
@@ -1,188 +1,0 @@
-const dickens = `
-It was the Dover road that lay, on a Friday night late in November, before the first of the persons with whom this history has business. The Dover road lay, as to him, beyond the Dover mail, as it lumbered up Shooter's Hill. He walked up hill in the mire by the side of the mail, as the rest of the passengers did; not because they had the least relish for walking exercise, under the circumstances, but because the hill, and the harness, and the mud, and the mail, were all so heavy, that the horses had three times already come to a stop, besides once drawing the coach across the road, with the mutinous intent of taking it back to Blackheath. Reins and whip and coachman and guard, however, in combination, had read that article of war which forbade a purpose otherwise strongly in favour of the argument, that some brute animals are endued with Reason; and the team had capitulated and returned to their duty.
-
-With drooping heads and tremulous tails, they mashed their way through the thick mud, floundering and stumbling between whiles, as if they were falling to pieces at the larger joints. As often as the driver rested them and brought them to a stand, with a wary “Wo-ho! so-ho-then!” the near leader violently shook his head and everything upon it—like an unusually emphatic horse, denying that the coach could be got up the hill. Whenever the leader made this rattle, the passenger started, as a nervous passenger might, and was disturbed in mind.
-
-There was a steaming mist in all the hollows, and it had roamed in its forlornness up the hill, like an evil spirit, seeking rest and finding none. A clammy and intensely cold mist, it made its slow way through the air in ripples that visibly followed and overspread one another, as the waves of an unwholesome sea might do. It was dense enough to shut out everything from the light of the coach-lamps but these its own workings, and a few yards of road; and the reek of the labouring horses steamed into it, as if they had made it all.
-
-Two other passengers, besides the one, were plodding up the hill by the side of the mail. All three were wrapped to the cheekbones and over the ears, and wore jack-boots. Not one of the three could have said, from anything he saw, what either of the other two was like; and each was hidden under almost as many wrappers from the eyes of the mind, as from the eyes of the body, of his two companions. In those days, travellers were very shy of being confidential on a short notice, for anybody on the road might be a robber or in league with robbers. As to the latter, when every posting-house and ale-house could produce somebody in “the Captain's” pay, ranging from the landlord to the lowest stable non-descript, it was the likeliest thing upon the cards. So the guard of the Dover mail thought to himself, that Friday night in November, one thousand seven hundred and seventy-five, lumbering up Shooter's Hill, as he stood on his own particular perch behind the mail, beating his feet, and keeping an eye and a hand on the arm-chest before him, where a loaded blunderbuss lay at the top of six or eight loaded horse-pistols, deposited on a substratum of cutlass.
-
-The Dover mail was in its usual genial position that the guard suspected the passengers, the passengers suspected one another and the guard, they all suspected everybody else, and the coachman was sure of nothing but the horses; as to which cattle he could with a clear conscience have taken his oath on the two Testaments that they were not fit for the journey.
-
-“Wo-ho!” said the coachman. “So, then! One more pull and you're at the top and be damned to you, for I have had trouble enough to get you to it!—Joe!”
-
-“Halloa!” the guard replied.
-
-“What o'clock do you make it, Joe?”
-
-“Ten minutes, good, past eleven.”
-
-“My blood!” ejaculated the vexed coachman, “and not atop of Shooter's yet! Tst! Yah! Get on with you!”
-
-The emphatic horse, cut short by the whip in a most decided negative, made a decided scramble for it, and the three other horses followed suit. Once more, the Dover mail struggled on, with the jack-boots of its passengers squashing along by its side. They had stopped when the coach stopped, and they kept close company with it. If any one of the three had had the hardihood to propose to another to walk on a little ahead into the mist and darkness, he would have put himself in a fair way of getting shot instantly as a highwayman.
-
-The last burst carried the mail to the summit of the hill. The horses stopped to breathe again, and the guard got down to skid the wheel for the descent, and open the coach-door to let the passengers in.
-
-“Tst! Joe!” cried the coachman in a warning voice, looking down from his box.
-
-“What do you say, Tom?”
-
-They both listened.
-
-“I say a horse at a canter coming up, Joe.”
-
-“I say a horse at a gallop, Tom,” returned the guard, leaving his hold of the door, and mounting nimbly to his place. “Gentlemen! In the king's name, all of you!”
-
-With this hurried adjuration, he cocked his blunderbuss, and stood on the offensive.
-
-The passenger booked by this history, was on the coach-step, getting in; the two other passengers were close behind him, and about to follow. He remained on the step, half in the coach and half out of; they remained in the road below him. They all looked from the coachman to the guard, and from the guard to the coachman, and listened. The coachman looked back and the guard looked back, and even the emphatic leader pricked up his ears and looked back, without contradicting.
-
-The stillness consequent on the cessation of the rumbling and labouring of the coach, added to the stillness of the night, made it very quiet indeed. The panting of the horses communicated a tremulous motion to the coach, as if it were in a state of agitation. The hearts of the passengers beat loud enough perhaps to be heard; but at any rate, the quiet pause was audibly expressive of people out of breath, and holding the breath, and having the pulses quickened by expectation.
-
-The sound of a horse at a gallop came fast and furiously up the hill.
-
-“So-ho!” the guard sang out, as loud as he could roar. “Yo there! Stand! I shall fire!”
-
-The pace was suddenly checked, and, with much splashing and floundering, a man's voice called from the mist, “Is that the Dover mail?”
-
-“Never you mind what it is!” the guard retorted. “What are you?”
-
-“Is that the Dover mail?”
-
-“Why do you want to know?”
-
-“I want a passenger, if it is.”
-
-“What passenger?”
-
-“Mr. Jarvis Lorry.”
-
-Our booked passenger showed in a moment that it was his name. The guard, the coachman, and the two other passengers eyed him distrustfully.
-
-“Keep where you are,” the guard called to the voice in the mist, “because, if I should make a mistake, it could never be set right in your lifetime. Gentleman of the name of Lorry answer straight.”
-
-“What is the matter?” asked the passenger, then, with mildly quavering speech. “Who wants me? Is it Jerry?”
-
-(“I don't like Jerry's voice, if it is Jerry,” growled the guard to himself. “He's hoarser than suits me, is Jerry.”)
-
-“Yes, Mr. Lorry.”
-
-“What is the matter?”
-
-“A despatch sent after you from over yonder. T. and Co.”
-
-“I know this messenger, guard,” said Mr. Lorry, getting down into the road—assisted from behind more swiftly than politely by the other two passengers, who immediately scrambled into the coach, shut the door, and pulled up the window. “He may come close; there's nothing wrong.”
-
-“I hope there ain't, but I can't make so 'Nation sure of that,” said the guard, in gruff soliloquy. “Hallo you!”
-
-“Well! And hallo you!” said Jerry, more hoarsely than before.
-
-“Come on at a footpace! d'ye mind me? And if you've got holsters to that saddle o' yourn, don't let me see your hand go nigh 'em. For I'm a devil at a quick mistake, and when I make one it takes the form of Lead. So now let's look at you.”
-
-The figures of a horse and rider came slowly through the eddying mist, and came to the side of the mail, where the passenger stood. The rider stooped, and, casting up his eyes at the guard, handed the passenger a small folded paper. The rider's horse was blown, and both horse and rider were covered with mud, from the hoofs of the horse to the hat of the man.
-
-“Guard!” said the passenger, in a tone of quiet business confidence.
-
-The watchful guard, with his right hand at the stock of his raised blunderbuss, his left at the barrel, and his eye on the horseman, answered curtly, “Sir.”
-
-“There is nothing to apprehend. I belong to Tellson's Bank. You must know Tellson's Bank in London. I am going to Paris on business. A crown to drink. I may read this?”
-
-“If so be as you're quick, sir.”
-
-He opened it in the light of the coach-lamp on that side, and read—first to himself and then aloud: “'Wait at Dover for Mam'selle.' It's not long, you see, guard. Jerry, say that my answer was, Recalled to life.”
-
-Jerry started in his saddle. “That's a Blazing strange answer, too,” said he, at his hoarsest.
-
-“Take that message back, and they will know that I received this, as well as if I wrote. Make the best of your way. Good night.”
-
-With those words the passenger opened the coach-door and got in; not at all assisted by his fellow-passengers, who had expeditiously secreted their watches and purses in their boots, and were now making a general pretence of being asleep. With no more definite purpose than to escape the hazard of originating any other kind of action.
-
-The coach lumbered on again, with heavier wreaths of mist closing round it as it began the descent. The guard soon replaced his blunderbuss in his arm-chest, and, having looked to the rest of its contents, and having looked to the supplementary pistols that he wore in his belt, looked to a smaller chest beneath his seat, in which there were a few smith's tools, a couple of torches, and a tinder-box. For he was furnished with that completeness that if the coach-lamps had been blown and stormed out, which did occasionally happen, he had only to shut himself up inside, keep the flint and steel sparks well off the straw, and get a light with tolerable safety and ease (if he were lucky) in five minutes.
-
-“Tom!” softly over the coach roof.
-
-“Hallo, Joe.”
-
-“Did you hear the message?”
-
-“I did, Joe.”
-
-“What did you make of it, Tom?”
-
-“Nothing at all, Joe.”
-
-“That's a coincidence, too,” the guard mused, “for I made the same of it myself.”
-
-Jerry, left alone in the mist and darkness, dismounted meanwhile, not only to ease his spent horse, but to wipe the mud from his face, and shake the wet out of his hat-brim, which might be capable of holding about half a gallon. After standing with the bridle over his heavily-splashed arm, until the wheels of the mail were no longer within hearing and the night was quite still again, he turned to walk down the hill.
-
-“After that there gallop from Temple Bar, old lady, I won't trust your fore-legs till I get you on the level,” said this hoarse messenger, glancing at his mare. “'Recalled to life.' That's a Blazing strange message. Much of that wouldn't do for you, Jerry! I say, Jerry! You'd be in a Blazing bad way, if recalling to life was to come into fashion, Jerry!”
-A wonderful fact to reflect upon, that every human creature is constituted to be that profound secret and mystery to every other. A solemn consideration, when I enter a great city by night, that every one of those darkly clustered houses encloses its own secret; that every room in every one of them encloses its own secret; that every beating heart in the hundreds of thousands of breasts there, is, in some of its imaginings, a secret to the heart nearest it! Something of the awfulness, even of Death itself, is referable to this. No more can I turn the leaves of this dear book that I loved, and vainly hope in time to read it all. No more can I look into the depths of this unfathomable water, wherein, as momentary lights glanced into it, I have had glimpses of buried treasure and other things submerged. It was appointed that the book should shut with a spring, for ever and for ever, when I had read but a page. It was appointed that the water should be locked in an eternal frost, when the light was playing on its surface, and I stood in ignorance on the shore. My friend is dead, my neighbour is dead, my love, the darling of my soul, is dead; it is the inexorable consolidation and perpetuation of the secret that was always in that individuality, and which I shall carry in mine to my life's end. In any of the burial-places of this city through which I pass, is there a sleeper more inscrutable than its busy inhabitants are, in their innermost personality, to me, or than I am to them?
-
-As to this, his natural and not to be alienated inheritance, the messenger on horseback had exactly the same possessions as the King, the first Minister of State, or the richest merchant in London. So with the three passengers shut up in the narrow compass of one lumbering old mail coach; they were mysteries to one another, as complete as if each had been in his own coach and six, or his own coach and sixty, with the breadth of a county between him and the next.
-
-The messenger rode back at an easy trot, stopping pretty often at ale-houses by the way to drink, but evincing a tendency to keep his own counsel, and to keep his hat cocked over his eyes. He had eyes that assorted very well with that decoration, being of a surface black, with no depth in the colour or form, and much too near together—as if they were afraid of being found out in something, singly, if they kept too far apart. They had a sinister expression, under an old cocked-hat like a three-cornered spittoon, and over a great muffler for the chin and throat, which descended nearly to the wearer's knees. When he stopped for drink, he moved this muffler with his left hand, only while he poured his liquor in with his right; as soon as that was done, he muffled again.
-
-“No, Jerry, no!” said the messenger, harping on one theme as he rode. “It wouldn't do for you, Jerry. Jerry, you honest tradesman, it wouldn't suit your line of business! Recalled—! Bust me if I don't think he'd been a drinking!”
-
-His message perplexed his mind to that degree that he was fain, several times, to take off his hat to scratch his head. Except on the crown, which was raggedly bald, he had stiff, black hair, standing jaggedly all over it, and growing down hill almost to his broad, blunt nose. It was so like Smith's work, so much more like the top of a strongly spiked wall than a head of hair, that the best of players at leap-frog might have declined him, as the most dangerous man in the world to go over.
-
-While he trotted back with the message he was to deliver to the night watchman in his box at the door of Tellson's Bank, by Temple Bar, who was to deliver it to greater authorities within, the shadows of the night took such shapes to him as arose out of the message, and took such shapes to the mare as arose out of her private topics of uneasiness. They seemed to be numerous, for she shied at every shadow on the road.
-
-What time, the mail-coach lumbered, jolted, rattled, and bumped upon its tedious way, with its three fellow-inscrutables inside. To whom, likewise, the shadows of the night revealed themselves, in the forms their dozing eyes and wandering thoughts suggested.
-
-Tellson's Bank had a run upon it in the mail. As the bank passenger—with an arm drawn through the leathern strap, which did what lay in it to keep him from pounding against the next passenger, and driving him into his corner, whenever the coach got a special jolt—nodded in his place, with half-shut eyes, the little coach-windows, and the coach-lamp dimly gleaming through them, and the bulky bundle of opposite passenger, became the bank, and did a great stroke of business. The rattle of the harness was the chink of money, and more drafts were honoured in five minutes than even Tellson's, with all its foreign and home connection, ever paid in thrice the time. Then the strong-rooms underground, at Tellson's, with such of their valuable stores and secrets as were known to the passenger (and it was not a little that he knew about them), opened before him, and he went in among them with the great keys and the feebly-burning candle, and found them safe, and strong, and sound, and still, just as he had last seen them.
-
-But, though the bank was almost always with him, and though the coach (in a confused way, like the presence of pain under an opiate) was always with him, there was another current of impression that never ceased to run, all through the night. He was on his way to dig some one out of a grave.
-
-Now, which of the multitude of faces that showed themselves before him was the true face of the buried person, the shadows of the night did not indicate; but they were all the faces of a man of five-and-forty by years, and they differed principally in the passions they expressed, and in the ghastliness of their worn and wasted state. Pride, contempt, defiance, stubbornness, submission, lamentation, succeeded one another; so did varieties of sunken cheek, cadaverous colour, emaciated hands and figures. But the face was in the main one face, and every head was prematurely white. A hundred times the dozing passenger inquired of this spectre:
-
-“Buried how long?”
-
-The answer was always the same: “Almost eighteen years.”
-
-“You had abandoned all hope of being dug out?”
-
-“Long ago.”
-
-“You know that you are recalled to life?”
-
-“They tell me so.”
-
-“I hope you care to live?”
-
-“I can't say.”
-
-“Shall I show her to you? Will you come and see her?”
-
-The answers to this question were various and contradictory. Sometimes the broken reply was, “Wait! It would kill me if I saw her too soon.” Sometimes, it was given in a tender rain of tears, and then it was, “Take me to her.” Sometimes it was staring and bewildered, and then it was, “I don't know her. I don't understand.”
-
-After such imaginary discourse, the passenger in his fancy would dig, and dig, dig—now with a spade, now with a great key, now with his hands—to dig this wretched creature out. Got out at last, with earth hanging about his face and hair, he would suddenly fan away to dust. The passenger would then start to himself, and lower the window, to get the reality of mist and rain on his cheek.
-
-Yet even when his eyes were opened on the mist and rain, on the moving patch of light from the lamps, and the hedge at the roadside retreating by jerks, the night shadows outside the coach would fall into the train of the night shadows within. The real Banking-house by Temple Bar, the real business of the past day, the real strong rooms, the real express sent after him, and the real message returned, would all be there. Out of the midst of them, the ghostly face would rise, and he would accost it again.
-
-“Buried how long?”
-
-“Almost eighteen years.”
-
-“I hope you care to live?”
-
-“I can't say.”
-
-Dig—dig—dig—until an impatient movement from one of the two passengers would admonish him to pull up the window, draw his arm securely through the leathern strap, and speculate upon the two slumbering forms, until his mind lost its hold of them, and they again slid away into the bank and the grave.
-
-“Buried how long?”
-
-“Almost eighteen years.”
-
-“You had abandoned all hope of being dug out?”
-
-“Long ago.”
-
-The words were still in his hearing as just spoken—distinctly in his hearing as ever spoken words had been in his life—when the weary passenger started to the consciousness of daylight, and found that the shadows of the night were gone.
-
-He lowered the window, and looked out at the rising sun. There was a ridge of ploughed land, with a plough upon it where it had been left last night when the horses were unyoked; beyond, a quiet coppice-wood, in which many leaves of burning red and golden yellow still remained upon the trees. Though the earth was cold and wet, the sky was clear, and the sun rose bright, placid, and beautiful.
-
-“Eighteen years!” said the passenger, looking at the sun. “Gracious Creator of day! To be buried alive for eighteen years!”
-`
-
-export default dickens;
